Removed a few compiler warnings.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@1443 01821946-851b-0410-b7f4-f1477c70e57a

+++ lighting/dynlight.cpp
@@ -787,9 +787,9 @@ activeLightsP [xDist].nType = nType;
 activeLightsP [xDist].psl = psl;
 gameData.render.lights.dynamic.shader.nActiveLights [nThread]++;
 if (gameData.render.lights.dynamic.shader.nFirstLight [nThread] > xDist)
-	gameData.render.lights.dynamic.shader.nFirstLight [nThread] = xDist;
+	gameData.render.lights.dynamic.shader.nFirstLight [nThread] = (short) xDist;
 if (gameData.render.lights.dynamic.shader.nLastLight [nThread] < xDist)
-	gameData.render.lights.dynamic.shader.nLastLight [nThread] = xDist;
+	gameData.render.lights.dynamic.shader.nLastLight [nThread] = (short) xDist;
 return 1;
 }
